+++ commands1/leavefor.js
@@ -15,7 +15,7 @@ module.exports.run = async (bot, msg, args) => {
   var mention = msg.mentions.users.first()
   user = await Data.findOne().byID(mention.id)
   channel = msg.channel.name
-
+  if(!user) return msg.channel.send("Please try using the play command first")
   if (user.ingame !== 1) return msg.channel.send("Cant leave a game if havent joined one!")
 
   bot.guilds.cache.get(config.SERVER_ID).channels.cache.get("711048304502374493").messages.fetch(user.message).then(m => m.delete()).catch(err => {
